backend/api/pyproject.toml
[project]
name = "HomeOffer--api"
version = "0.1.0"
description = "API for HomeOffer"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dotenv>=0.9.9",
    "fastapi>=0.120.2",
    "faststream[rabbit]>=0.6.2",
    "pydantic-config>=0.3.0",
    "uvicorn>=0.38.0",
]
